DON-1003: Refactor: Merge mandate service and regular giving service classes
